+++ hata/__init__.py
@@ -107,7 +107,7 @@ interrupt.
 
 We got some tutorials on `github:https://github.com/HuyaneMatsu/hata/tree/master/docs` as well, please check them too!
 """
-__version__ = '1.2.11'
+__version__ = '1.2.12'
 
 from .utils.module_deprecation import get_deprecation_function